revert navigation block changes
